requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
Django==5.2.12
drf-spectacular==0.29.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
google-auth==2.49.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
requests==2.32.5
sqlparse==0.5.5
Pillow==11.1.0
